+++ CHANGELOG.md
@@ -5,7 +5,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [0.2.2](main)
 
-- ...
+- feature: SLURM support
+- feature: Support passing of extra arguments from CLI -> Snakemake for run commands
 
 ## [0.2.1](https://github.com/omnibenchmark/omnibenchmark/releases/tag/v0.2.1) (May 20th 2025)